vad : map timestamps to original audio
This commit adds a mapping of the original audio timestamps to the
timestamps of the segments in the VAD (Voice Activity Detection)
process.

The motivation for this change is when we process the original audio
signal and only pass the speech segments to whisper_full, the
timestamps that whisper returnes when calling functions like
whisper_full_get_segment_t0 are the timestamps for the "VAD"
segments and not the original audio.

The values are not identical to the the timestamps processed without VAD
enabled but they are close, and hopefully close enough.