vad : use orig timestamp for first segment
This commit removes the handling of the first segment which ensured that
it always started at zero. This was not correct and we should have been
showing the original timestamp that was stored by the VAD processing.

There might be better ways to handle this alighment of timestamps
perhaps should investigate this further.