vad : add probs to whisper_vad_state
This commit adds a `probs` member to the `whisper_vad_state` struct,
which is a vector of floats.

A pointer to this data is returned in the `whisper_vad_speech` struct
instead of copying the data to a new array, to the data will only be
valid for as long as the whisper_vad_state object exists. I was
concerned about the performance of copying the data, but I'm not sure if
this is a good way to handle this and perhaps a copy would be better.