whisper : add support for for skipped audio
This commit adds a check for skipped audio samples at the beginning of a
audio sample.

The motivation for this is that some models like medium which have been
trained to skip introduction/credits at the beginning of audio books.
This caused the segment start timestamp to start at 0 but the end
timestamp was relative to the skipped audio (so must larger and
incorrect). With this change the segment start timestamp is aligned with
the skipped audio.

Resolves: https://github.com/ggml-org/whisper.cpp/issues/3207