examples : remove unused code from common.{cpp,h}
This commit removes unused structs and functions from common.h and
common.cpp.

The motivation for this change to clean up the examples codebase and
this is a start to that process. I was going to add VAD support to the
server and actually thought of adding these params to common.h which is
why I started looking at this.