mcp : add initial MCP server example (wip)
This commit adds an suggestion for an Model Context Protocol (MCP) server
implementation.

There is a simple command line demo which uses the stdio transport
protcol (there is no http transport at this point) and its usage is
described in the [examples readme](examples/mcp/README.md).
The readme also contains an example of configuring Claude.ai Desktop to use
this MCP server and to transcribe (please see the
[Claude.ai Desktop integration section](examples/mcp/README.md#claudeai-desktop-integration)
for further details and screen shots).

The motivation for this was mainly that I wanted to learn more about MCP
and trying to integrate whisper.cpp was a good way of learning. I'm not
sure if this is something that we want to persue further, but I'd be
happy to do so if there is interest. There are missing features of the
protocol and also adding http transport (which should hopefully only be a
matter of implementing a mcp::Transport for http).